refactor: Extract InitialsAvatar
