+++ apps/web/src/features/organizations/components/InitialsAvatar/styles.module.css
@@ -1,0 +1,10 @@
+.initialsAvatar {
+    grid-area: logo;
+    text-transform: uppercase;
+    font-weight: bold;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: white;
+    justify-self: start;
+}
