+++ apps/web/src/features/organizations/components/OrgsCard/index.tsx
@@ -1,5 +1,5 @@
 import { AppRoutes } from '@/config/routes'
-import { Box, Card, hslToRgb, Stack, Typography } from '@mui/material'
+import { Box, Card, Stack, Typography } from '@mui/material'
 import MoreVertIcon from '@mui/icons-material/MoreVert'
 import IconButton from '@mui/material/IconButton'
 import Link from 'next/link'
@@ -8,53 +8,7 @@ import css from './styles.module.css'
 import type { GetOrganizationResponse } from '@safe-global/store/gateway/AUTO_GENERATED/organizations'
 import classNames from 'classnames'
 import { useOrgSafeCount } from '@/features/organizations/hooks/useOrgSafeCount'
-
-/**
- * Returns a deterministic "random" color (in Hex format) based on a string.
- * The color is constrained so it won't be too dark or too light or too saturated.
- */
-export function getDeterministicColor(str: string): string {
-  const sum = [...str].reduce((acc, char) => acc + char.charCodeAt(0), 0)
-
-  const hue = sum % 360
-  const saturation = 40 + (sum % 31)
-  const lightness = 40 + (sum % 31)
-
-  return hslToRgb(`hsl(${hue}, ${saturation}, ${lightness})`)
-}
-
-export const InitialsAvatar = ({
-  name,
-  size = 'large',
-  rounded = false,
-}: {
-  name: string
-  size?: 'small' | 'medium' | 'large'
-  rounded?: boolean
-}) => {
-  const logoLetters = name.slice(0, 1)
-  const logoColor = getDeterministicColor(name)
-  const dimensions = {
-    small: { width: 24, height: 24, fontSize: '12px !important' },
-    medium: { width: 32, height: 32, fontSize: '16px !important' },
-    large: { width: 48, height: 48, fontSize: '20px !important' },
-  }
-
-  const { width, height, fontSize } = dimensions[size]
-
-  return (
-    <Box
-      className={css.initialsAvatar}
-      bgcolor={logoColor}
-      width={width}
-      height={height}
-      fontSize={fontSize}
-      borderRadius={rounded ? '50%' : '6px'}
-    >
-      {logoLetters}
-    </Box>
-  )
-}
+import InitialsAvatar from '@/features/organizations/components/InitialsAvatar'
 
 export const OrgSummary = ({
   name,
