+++ apps/web/src/features/organizations/components/OrgsCard/styles.module.css
@@ -20,17 +20,6 @@
   border: 0;
 }
 
-.initialsAvatar {
-  grid-area: logo;
-  text-transform: uppercase;
-  font-weight: bold;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: white;
-  justify-self: start;
-}
-
 .orgInfo {
   grid-area: info;
 }
